+++ lib/src/calendar/cupertino_calendar.dart
@@ -128,6 +128,7 @@ class CupertinoCalendar extends StatefulWidget {
   /// For 24h format being used or not, results in AM/PM being shown or hidden in the widget.
   /// Setting to `true` or `false` will force 24h format to be on or off.
   /// The default value is null, which calls [MediaQuery.alwaysUse24HourFormatOf].
+  ///
   /// Displayed only when the calendar is in a mode that includes time selection.
   final bool? use24hFormat;