+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 2.0.1
 
 * `use24hFormat` param has been added to all widgets.
-* `dismissBehavior` param has been added to the `CupertinoCalendarPickerButton`.
+* `dismissBehavior` param has been added to the `CupertinoCalendarPickerButton` widget.
 
 ## 2.0.0+1